arcs/lambda/app/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx==0.27.2
cachetools==5.5.0
PyNaCl==1.5.0
